+++ src/main/java/step3/view/ReturnView.java
@@ -1,7 +1,7 @@
 package step3.view;
 
 import step3.LottoPapers;
-import step3.Winner;
+import step3.winner.Winner;
 
 public class ReturnView implements ResultView {
